Fixed the WORST BUG known to MANKIND
# Uncovering the mystery
`Enemy.health`, instead of being a normal decimal like 0.45 (so the player does .15 damage of an enemy with a health of .6), chooses to look something like this: `0.4499999999999999` or this `0.450000001`. So when the enemy is killed, its health is ACTUALLY `0.00000000001`. But the Godot engine chooses to ROUND the number when I call `print()`. ROUNDING THE NUMBER. I spent *countless* hours trying to find out what was making this happen and I finally found out! (Well, most of those hours were spent just living with the bug because I was too lazy to deal with it). At first I thought it was a GDScript bug. Well...