Fixed a bug with ammo powerups (see description for more)
The error in question occurred only if the player had no more ammo and ammo refills. If the player was then to pick up an ammo powerup and press the spacebar to start a reload, it wouldn't work. So in short, after you expended all your ammo and all your refills, you couldn't get any more.