"Redesigned" the settings screen to cover up a larger issue with the selection square
The issue: The select square draws on TOP of other stuff, so nodes that would be invisible because of a ScrollContainer node would still have a select square drawn on them. With the previous code, that would result in jittery movement when the ScrollContainer node is scrolled. This new design may look worse, but I fixed the problem by just moving the select square *inside* the ScrollContainer node (so it would be culled too). Then I had to disable clipping for the ScrollContainer node so all of the square can be visible and now we have this.