Made some loose code to make the docking station arrive after each level
NOTE 1: I also, finding bosses too difficult, lowered the damage of all bosses by about a half

NOTE 2: My plan is to have upgrades persist only per-run as to fully remove that feeling of being too powerful for the first level. So instead, the goal is to display the upgrades screen as part of a thing that happens when you board a ship after each level. It looks a bit clunky now, but I am to fix that in the future.

NOTE 3: So note-to-self here: the upgrades screen still makes upgrades persist with the userdata.txt file. I'll need to edit it so it only edits scores/userdata per run. Also, I'll need to make the experience a bit more smooth.