Followed a tutorial to try to get some scenes to load faster
Basically, as I understand it, when scenes like lasers and enemy ships (which contain Explosion scene instances) are created for the first time, Godot has to compile ParticlesMaterial shaders for the various particle effects these nodes have. So what the player experiences is some lag when they shoot a laser for the first time or an enemy spawns for the first time, which can be quite annoying. What this should do is compile these shaders *before* a game starts by auto-loading Particles instances, forcing Godot to compile and cache the shaders when the game starts. Very cool.