Some more performance improvements
And with that I think I've done as much as I can. There's still a problem where instancing scenes for the first time will cause quite a bit of lag on some computers, but I searched it up and it looks like something I can't fix. So that's all I can do for InfiniteShooter 4. :(
Maybe I'll port InfiniteShooter to Godot 4 for the next release of the game, though!