Made a (heavily WIP) loading screen.
This *might* work? It works like the ParticlesCache node, forcing Godot to instance all scenes first. (My logic is that since there are lag spikes when instancing scenes for the first time we can just instance EVERY scene for the first time under a loading screen so users know to expect lag)