Set the screen size and almost completely implemented walls
NOTE:

I decided to change the screen size to landscape (but still 4:3). I think this will make the game a bit more engaging and allow more breathing room. The walls will then be used to ensure the same size for the "play-space" is shown no matter the size.

I set the perspective camera so that when the game window is resized it will preserve the height. This doesn't have any effect on the GUI though, but what ends up happening is that the width of what the camera can see increases. If I want to have a certain size for this "play-space" (because it affects the difficulty of the game), I need to instill some bounds on it. My best idea so far is to have some walls so that it looks nice.

That leads me to where I am now. I finally found a way to get it working, but after writing a lot of spaghetti code. It *works*, but I need to simplify it, find out what the hell I just did, and then comment my code. Then I can add some final touches and polish to this new system and I'm done!