(Should've) stopped bosses from cutting off particles emitted by LaserEffects while they die.
Fingers crossed.
A bit more description for the issue: If you light a boss on fire, and it dies whilst still on fire, the particles for it burning will just abruptly cut off. Plus, it would still burn even though the actual ship just exploded and thus cannot burn [citation needed].

Oh, and I also got red of a couple errors in the debugger.