Reworked resizing
I understand this may be buggy, but it's the best I can do given what I have.
It's supposed to maintain an aspect ratio of 4:3 unless the window is wider than that, in which case the walls will show. But I'm having some issues -- quite a lot, in fact -- trying to get the GUI to listen to me and scale as well. Eventually, after headaches and billions of tabs open, I decided to just give up. I am sorry. :(