Replaced usernames in the leaderboard for dates
InfiniteShooter's a game that'll be installed locally to a computer, so it's more useful to see *when* a score was achieved rather than who scored it (which will always be yourself). Plus, it should actually provide some incentive to beat your previous scores since you know when you scored them.