(Hopefully) Last commit before 4.0!
After staying up until 1:00 in the morning testing InfiniteShooter, I can confirm that it is playable, balanced, and for the most part bug-free. I know the debugger still gives out some warnings, but I think it's fine. The game just works(tm).
But I think the main reason why I want this to be the last commit before 4.0 is that I'm just really tired. I worked on InfiniteShooter a *lot* and I can't just set it down to take a break, then pick it up a month or two later. I need to make sure the game is complete, and then iron out some bugs in future releases, because I just need to ship a *version*.
But I'm still a perfectionist, so I tested the game to make sure you could comfortably get to at least level 6. The game introduces new enemies past that level, but it's a good mark for when the game is playable. I might have to do a bit more balancing as I get user feedback (assuming I would), but I think--especially since I'm the only playtester--this is done. And long overdue too.
So that's it! That's InfiniteShooter 4! Now, to release it...
Oh, and if I find any more bugs before then, I guess I'll have to do some more tweaking and playtesting...