The (actual) final commit
Well. I'm a perfectionist.

I fixed a bit of bugs, did some last minute polish, and added a new mechanic to balance the game (and fully playtested it to ensure it worked): Coin crates.
Every level, a set amount of coin crates spawn. If you manage to shoot and hit them (they're spinning, so that makes it rather difficult to do) you will be rewarded with coins you can use for upgrades. And that's about it!

I planned to write more here, but I'm *reallly* tired. And I just finished ALL of InfiniteShooter 4 to the point where I can say that I'm actually proud of it. So that's it! The only next thing to do is to publish my game!
Well, I don't know what you're doing, but it's midnight so I'm going to sleep. Have a good night everyone! Thanks, and have fun with InfiniteShooter 4.0!