0.4.1: Remove the option to set the keybinding to slow time
WELL. Essentially, it was a mechanic I had a while ago, but I found it to be too overpowered without making significant gameplay changes so I removed it. For some reason, I didn't even notice that the option to set the key for it was still there, and that made it to InfiniteShooter 0.4.0. Oh well. I removed it now.