0.4.1: Some small changes to boss mechanics
I think it should be enough to do this for 0.4.1 and not 0.5.0, but I made some small changes to boss mechanics that will depend on your difficulty. It's not perfect--that's what I'm hoping to strive for in 0.5.0, but this should at least give some variety for different difficulties so they can actually *be* more/less difficult.