Disable physics interpolation by default
At the heart of the stuttering is physics interpolation; objects that are being teleported (and are supposed to teleport) are being interpolated by the engine in split-second events causing them to stutter. For instance, the walls are supposed to look infinite by teleporting to the top of the screen but end up sliding up (just really, really fast) because the interpolation.

I was initially planning on making a setting for this, but since there's already some edits to the UI for 0.5.0 and I don't want new changes to conflict with what's already been done I'll add it in `feature-additions` instead.