Add mouse controls to the player
I had to rework what I used to have for mouse controls (see the previous commit for that) because it was ending up with movement that was too slow. Plus, there was still jank when I used my TrackPoint (because of unexplained engine reasons???), so I spent the last while rewriting this code to make sure it worked smoothly. Code should be a lot easier to read now and hopefully the controls will feel nice when you're using the mouse.

It's actually easier than you think (and I got the epiphany for it while playing Doom and thinking about what it would have had for mouse controls): actions with mouse controls are only done when the mouse *moves*, and it's not interpolated because it doesn't need to be. Take a look at [OpenTyrian's controls](https://github.com/opentyrian/opentyrian/blob/master/src/keyboard.c): it's using SDL to say that when the mouse is moved, map its coordinates to a position on the screen and do whatever is usually done for controls there. So as it turns out, all I needed to do was write code that moves the player with a mouse only in an event when the mouse was moved, and as it turned out that actually ended up solving the issues I was having with controls.