Update the HUD to show when a boss fight is happening
Although this fits better with issue #25, it's needed here because before a boss spawns you actually, technically, get to wave 6. So when you're in a boss fight on level 1 you'll see the HUD say that you're still on wave 5, but save and come back to the game and you'll find it saying you're on wave 6.